requirements.txt
streamlit==1.37.0
pydantic==2.8.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
SQLAlchemy==2.0.32
pandas==2.2.2
numpy<2.0,>=1.23.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.43.3
scikit-learn==1.5.1
python-dotenv==1.0.1
jinja2==3.1.4
reportlab==4.2.2
PyMySQL==1.1.1
sentencepiece 
protobuf
